# openapi.yaml
...
openapi: 3.1.0
info:
  title: aperIA - Python API
  description: '

    API do **aperIA** — ASPM (Application Security Posture Management) open-source.


    Um webhook de Pull Request do GitHub dispara um pipeline de scanners de

    segurança em 3 tiers (via Celery). O Claude raciocina como atacante

    (correlaciona findings, monta attack paths) e entrega patches como

    **GitHub code suggestions** para aprovação humana.


    > **Premissa inviolável:** o aperIA **nunca** aplica código sozinho.

    > Todo patch sai como code suggestion para aprovação humana.

    '
  contact:
    name: aperIA
  license:
    name: Open Source
  version: 1.0.0
paths:
...
  /findings:
    get:
      tags:
      - findings
      summary: Listar findings
      description: 'Lista findings do USUÁRIO logado, com filtros opcionais e paginacao.


        Isolado por dono: só retorna findings de commits cujos scans pertencem ao

        usuário autenticado. Ordenados do mais recente para o mais antigo; o campo

        `raw_output` é omitido aqui — use `GET /findings/{finding_id}` para o detalhe.'
      operationId: list_findings_findings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: commit_sha
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtra pelo SHA do commit.
          title: Commit Sha
        description: Filtra pelo SHA do commit.
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FindingSeverity'
          - type: 'null'
          description: Filtra pela severidade.
          title: Severity
        description: Filtra pela severidade.
      - name: tier
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 3
            minimum: 1
          - type: 'null'
          description: Filtra pelo tier (1-3).
          title: Tier
        description: Filtra pelo tier (1-3).
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filtra pela ferramenta de origem (ex.: semgrep).'
          title: Source
        description: 'Filtra pela ferramenta de origem (ex.: semgrep).'
      - name: secret_verified
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filtra secrets verificados.
          title: Secret Verified
        description: Filtra secrets verificados.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Tamanho da pagina.
          default: 50
          title: Limit
        description: Tamanho da pagina.
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Deslocamento para paginacao.
          default: 0
          title: Offset
        description: Deslocamento para paginacao.
      responses:
        '200':
          description: Pagina de findings que atendem aos filtros.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingPage'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /findings/{finding_id}:
    get:
      tags:
      - findings
      summary: Consultar finding por ID
      description: 'Retorna o detalhe de um finding do usuário. 404 se não existir ou não

        pertencer a um scan do usuário autenticado (não vaza existência).'
      operationId: get_finding_findings__finding_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: finding_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Finding Id
      responses:
        '200':
          description: Detalhe do finding, incluindo `raw_output`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingDetail'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Finding nao encontrado.
          content:
            application/json:
              example:
                detail: Finding nao encontrado
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /repositories/{repository_id}/findings:
    get:
      tags:
      - repositories
      summary: Listar findings do repositório
      description: Lista os findings do repositório (todos os commits). 404 se não for do usuário.
      operationId: list_repository_findings_repositories__repository_id__findings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FindingSeverity'
          - type: 'null'
          title: Severity
      - name: tier
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 3
            minimum: 1
          - type: 'null'
          title: Tier
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingPage'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FindingDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        source:
          type: string
          title: Source
        severity:
          $ref: '#/components/schemas/FindingSeverity'
        tier:
          type: integer
          title: Tier
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        cve_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cve Id
        cwe_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
        file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: File Path
        line_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Line Number
        asset:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset
        asset_criticality:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Criticality
        secret_verified:
          type: boolean
          title: Secret Verified
        secret_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Type
        commit_sha:
          type: string
          title: Commit Sha
        repo_url:
          type: string
          title: Repo Url
        created_at:
          type: string
          format: date-time
          title: Created At
        raw_output:
          additionalProperties: true
          type: object
          title: Raw Output
      type: object
      required:
      - id
      - source
      - severity
      - tier
      - title
      - description
      - cve_id
      - cwe_id
      - file_path
      - line_number
      - asset
      - asset_criticality
      - secret_verified
      - secret_type
      - commit_sha
      - repo_url
      - created_at
      - raw_output
      title: FindingDetail
      description: Representacao detalhada de um finding, incluindo `raw_output` do scanner.
    FindingPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FindingResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: FindingPage
      description: Pagina de resultados para listagem de findings.
    FindingResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        source:
          type: string
          title: Source
        severity:
          $ref: '#/components/schemas/FindingSeverity'
        tier:
          type: integer
          title: Tier
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        cve_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cve Id
        cwe_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
        file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: File Path
        line_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Line Number
        asset:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset
        asset_criticality:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Criticality
        secret_verified:
          type: boolean
          title: Secret Verified
        secret_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Type
        commit_sha:
          type: string
          title: Commit Sha
        repo_url:
          type: string
          title: Repo Url
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - source
      - severity
      - tier
      - title
      - description
      - cve_id
      - cwe_id
      - file_path
      - line_number
      - asset
      - asset_criticality
      - secret_verified
      - secret_type
      - commit_sha
      - repo_url
      - created_at
      title: FindingResponse
      description: Representacao publica de um finding para listagem (sem `raw_output`).
    FindingSeverity:
      type: string
      enum:
      - critical
      - high
      - medium
      - low
      - info
      title: FindingSeverity
      description: Espelha o enum `Severity` do dominio para uso na camada de apresentacao.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...